script/python/gitAllBranchCommit/requirements.txt
GitPython==3.1.36
GitPython==3.1.40
